moved db templates
