+++ src/db/schema/templates/DynamoDbTableCreate.ts
