+++ src/db/schema/templates/empty_schema.sql
